chore: add smoke test and disable some before fns for now
