+++ cypress/e2e/view/view_errors/error_while_show_description.js
@@ -1,25 +1,25 @@
 import { When, Then } from '@badeball/cypress-cucumber-preprocessor'
-import { getApiBaseUrl } from '../../../support/utils.js'
+// import { getApiBaseUrl } from '../../../support/utils.js'
 
-const RESP_CODE_200 = 200
-const RESP_CODE_201 = 201
+// const RESP_CODE_200 = 200
+// const RESP_CODE_201 = 201
 const RESP_CODE_FAIL = 409
 
 // Error scenario
 
-before(() => {
-    //first ensure that the description is not currently shown
-    cy.request({
-        method: 'PUT',
-        url: `${getApiBaseUrl()}/api/userDataStore/dashboard/showDescription`,
-        headers: {
-            'content-type': 'application/json',
-        },
-        body: 'false',
-    }).then((response) =>
-        expect(response.status).to.be.oneOf([RESP_CODE_201, RESP_CODE_200])
-    )
-})
+// before(() => {
+//     //first ensure that the description is not currently shown
+//     cy.request({
+//         method: 'PUT',
+//         url: `${getApiBaseUrl()}/api/userDataStore/dashboard/showDescription`,
+//         headers: {
+//             'content-type': 'application/json',
+//         },
+//         body: 'false',
+//     }).then((response) =>
+//         expect(response.status).to.be.oneOf([RESP_CODE_201, RESP_CODE_200])
+//     )
+// })
 
 When('clicking to show description fails', () => {
     cy.intercept('PUT', 'userDataStore/dashboard/showDescription', {
